requirements.txt
# ============================================
# REQUIREMENTS - Extratos Contábeis API
# ============================================
# Python >= 3.11 required

# ============================================
# FastAPI e Servidor Web
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# ============================================
# Configuração e Validação
# ============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# Banco de Dados
# ============================================
sqlalchemy>=2.0.0
pymysql>=1.1.0

# ============================================
# Processamento de PDF
# ============================================
pypdf>=4.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# ============================================
# Integração LLM (OpenAI)
# ============================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.0.0

# ============================================
# Manipulação de Excel
# ============================================
openpyxl>=3.1.0
pandas>=2.1.0

# ============================================
# Matching de Texto (Similaridade)
# ============================================
rapidfuzz>=3.6.0

# ============================================
# Logging
# ============================================
structlog>=24.1.0

# ============================================
# Utilitários
# ============================================
python-dotenv>=1.0.0
aiofiles>=23.0.0

# ============================================
# Desenvolvimento (opcional)
# ============================================
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# httpx>=0.26.0
